Hi Aamir
Interesting.  Solution seems quite clever, but a bit of a hack, in that 
I don't immediately see a natural relationship between class loader and 
process rank.

Another possibility may be to attach a distinct ThreadGroup to each MPI 
"process".  So far as I can tell thread group, like context class 
loader, is inherited from parent thread by default.  To me it is 
somewhat more intuitive to associate process rank with a thread group 
than with a class loader.

Said ThreadGroup could similarly be used as a key into a hash table, but 
a more elegant solution (if it works) may be to define a subclass 
SMPDevProcess of ThreadGroup that has a `getID()' method.  Then to 
find 
rank you do something like:

 ((SMPDevProcess) Thread.getCurrentThread().getThreadGroup()).getID()

Initialized nprocs to available number of cores in the MPJRun class.

Modified init() and finish() methods of SMPDeviceImpl class to handle
new strategy of getting ids.



git-svn-id: svn://svn.code.sf.net/p/mpjexpress/code@70 8344c875-6720-0410-8839-cff773666e85