Rizwan's modification seem to be working OK. He modifed the code to get rid of the Jetty dependency.
So from the code it seems that you write a single string that contains contents of the configuration file. And then these are passed to daemons that physically write the file and the URL (for the file) is passed to NIO device. I think the idea is right and should work ...



git-svn-id: svn://svn.code.sf.net/p/mpjexpress/code@123 8344c875-6720-0410-8839-cff773666e85